Refactor volume package
