+++ pkg/cli/volume_types.go
@@ -1,4 +1,4 @@
-package volume
+package cli
 
 import "strings"
